requirements.txt
flask
flask_restful
flask_pymongo
python-dotenv
flask_bcrypt
flask_jwt_extended
datetime
flask_cors